CHROMIUM: ARM64: dts: rockchip: add thermal zone node for rk3399 SoCs
This adds thermal zone node to rk3399 dtsi, rk3399 thermal data is
including the cpu and gpu sensor zone node.

The thermal zone node is the node containing all the required info
for describing a thermal zone, including its cooling device bindings.
The thermal zone node must contain, apart from its own properties, one
sub-node containing trip nodes and one sub-node containing all the zone
cooling maps.

The following is the parameter is introduced:
* polling-delay:
The maximum number of milliseconds to wait between polls

* polling-delay-passive:
The maximum number of milliseconds to wait between polls when performing
passive cooling.

* trips:
A sub-node which is a container of only trip point nodes required to
describe the thermal zone.

* cooling-maps:
A sub-node which is a container of only cooling device map nodes, used to
describe the relation between trips and cooling devices.

* cooling-device:
A phandle of a cooling device with its specifier, referring to which
cooling device is used in this cooling specifier binding. In the cooling
specifier, the first cell is the minimum cooling state and the second cell
is the maximum cooling state used in this map.

* sustainable-power
The sustainable-power describes how much power a thermal zone can
dissipate at the desired control temperature. In the binding doc [0], it actually
gives some examples:

For reference, the sustainable power of a 4'' phone is typically 2000mW,
while on a 10'' tablet is around 4500mW.

Since the gru  is > 4'' tablet, we choose 2600 mW.

[0] Documentation/devicetree/bindings/thermal/thermal.txt

* contribution
According to [1], the contribution field for each cooling-device express
their relative power efficiency. Higher weights express higher power
efficiency.  Weighting is relative such that if each cooling device has a
weight of 1 they are considered equal. This is particularly useful in
heterogeneous systems where two cooling devices may perform the same kind
of compute, but with different efficiency.
The suggests value:
little cpus: 2048
big cpus: 1024
gpu: 2048
That's taking the save power policy.

[1] Documentation/thermal/power_allocator.txt

BUG=chrome-os-partner:50756
TEST=Build and boot on rk3399 gru, run fishtank and head loading to watch
the cpu frequency and performance, and the power.

Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Change-Id: Id521570febf87ef704e673fbb51387f4de484e36
Reviewed-on: https://chromium-review.googlesource.com/346369
Reviewed-by: Javi Merino <javi.merino@arm.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>