FROMLIST: regulator: pwm: Retrieve correct voltage
The continuous PWM voltage regulator is caching the voltage value in
the ->volt_uV field. While most of the time this value should reflect the
real voltage, sometime it can be sightly different if the PWM device
rounded the set_duty_cycle request.
Moreover, this value is not valid until someone has modified the regulator
output.

Remove the ->volt_uV field and always rely on the PWM state to calculate
the regulator output.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
(am from https://patchwork.kernel.org/patch/9152171/)

BUG=chrome-os-partner:52666
TEST=build and boot on kevin; play with cpufreq userspace governor

Change-Id: If7d615dfe684706da463fdc4e3bc3ad11e5d93cc
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348047
Reviewed-by: Douglas Anderson <dianders@chromium.org>