UPSTREAM: ASoC: Intel: Skylake: Create Pipe to widget list in soc probe
We need to Identify the DSP pipe type and based on it being a pass
thru pipeline or not, we need to copy the pipeline params.

Pipe to widget mapping was earlier done in pre PMD widget handler,
but since the pipe type would now be required in hw_params for
bypass pipelines we need to move this to be done during the ASoC
probe of the platform component.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
tag: next-20160614
(cherry picked from commit 287af4f9f2679fd897e492338e4729c68f0c4a17)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:53967
TEST=with the whole series applied, Audio works on Amenia.

Change-Id: I34811c38fa26e63394982b53d40c10a657237a17
Reviewed-on: https://chromium-review.googlesource.com/354145
Commit-Ready: Vedang Patel <vedang.patel@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>