FROMLIST: Documentation: mmc: sdhci-of-arasan: Add ability to export card clock
Some SD/eMMC PHYs (like the PHY from Arasan that is designed to work
with arasan,sdhci-5.1) need to know the card clock frequency in order to
function properly.  Physically in a SoC this clock is exported from the
SDHCI IP block to the PHY IP block and the PHY needs to know the speed.
Let's export the SDHCI card clock using a standard device tree mechanism
so that the PHY can get access to it and query the card clock frequency.

BUG=chrome-os-partner:53445
TEST=Stress eMMC

Change-Id: Ic22e2b887c1e05384e3e8c66fb547ec5b18961c1
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Rob Herring <robh@kernel.org>
(am from https://patchwork.kernel.org/patch/9174691/)
Reviewed-on: https://chromium-review.googlesource.com/350139