UPSTREAM: Revert "ecryptfs: forbid opening files without mmap handler"
This reverts commit 2f36db71009304b3f0b95afacd8eba1f9f046b87.

It fixed a local root exploit but also introduced a dependency on
the lower file system implementing an mmap operation just to open a file,
which is a bit of a heavy hammer.  The right fix is to have mmap depend
on the existence of the mmap handler instead.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
(cherry picked from commit 78c4e172412de5d0456dc00d2b34050aa0b683b5)
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>

BUG=chrome-os-partner:55277
TEST=Play store works

Change-Id: I2975a687ab49a386517ba0841d633c8d5b4db74d
Reviewed-on: https://chromium-review.googlesource.com/362713
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Guenter Roeck <groeck@google.com>