UPSTREAM: drivers/perf: arm-pmu: Handle per-interrupt affinity mask
On a big-little system, PMUs can be wired to CPUs using per CPU
interrups (PPI). In this case, it is important to make sure that
the enable/disable do happen on the right set of CPUs.

So instead of relying on the interrupt-affinity property, we can
use the actual percpu affinity that DT exposes as part of the
interrupt specifier. The DT binding is also updated to reflect
the fact that the interrupt-affinity property shouldn't be used
in that case.

BUG=chrome-os-partner:54889
TEST=Build and boot on kevin; Perf list to monitor.

Change-Id: I90ea3a646a58daf2a900a4e1190c2c7abe03e146
Acked-by: Rob Herring <robh@kernel.org>
Tested-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
(cherry picked from commit 19a469a58720ea96b649b06fb09ddfd3e831aa69)
Reviewed-on: https://chromium-review.googlesource.com/358318
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>