CHROMIUM: cros_ec: add __packed to prevent padding
While the particular usage in question is likely safe (struct
cros_ec_command is 32-bit aligned, followed by <=32-bit fields), it's
been suggested this is not a great pattern to follow for the general
case -- for example, if we follow a 'struct cros_ec_command' (which is
32-bit- but not 64-bit-aligned) with a struct that starts with a 64-bit
type (e.g., u64), the compiler may add padding.

Let's add __packed, to inform the compiler of our true intention -- to
have no padding between these struct elements.

BUG=none
TEST=keyboard and other EC events

Change-Id: Ic7774e689e1fbd3e9040c2ac0424eae36b4c8ad6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/363500
Reviewed-by: Guenter Roeck <groeck@chromium.org>